STLXML2EBU-TT: FIXED: Consider also TCP when applying timecode offset
When the parameters offsetInSeconds/offsetInFrames are used to specify the used
offset within the STLXML input file, that offset is subtracted from the TCI/TCO
values when the EBU-TT file is written.

While the TCP relates to the TCI/TCO values, up to now it has not been modified
during the conversion. This is fixed now.

Fixes #32.