STLXML2EBU-TT: FIX EXSLT/XPATH 2.0 FUNCTIONS
Some functions that require XPath 2.0 were used by mistake. As some
XSLT processors support XSLT 2.0 but not the necessary EXSLT functions,
the respective code locations have been modified to use either an EXSLT
or an XPath 2.0 solution by checking for the available functions.

If the needed functions of neither EXSLT nor XPath 2.0 are available,
the conversion terminates with an error message.

In addition, the EXSLT Common namespaces prefix was harmonised and
removed from the transformation result.

Update the general and the STLXML2EBU-TT README according to the
changed requirements.

Fixes #46.