Changes behavior of EndBox element
In a previous commit the STLXML2EBU-TT.xslt was chamged, such that
the subtitle boxes in STLXML (signaled by elements StartBox and
EndBox) do not create new span elements in the resulting EBU-TT
file. This commit reverses these changes. Boxes do create new span
elements.

The schematron files for the corresponding tests were updated
accordingly.

The test file for the StartBox element was modifies such that two
StartBox and EndBox elements are used for each box, as it is
common in STL files.