Changes type of variable 'boxStarted'
The type of the variable 'boxStarted' is changed to boolean. It was
of type string ('used with string 'yes' and 'no') due to the XPath
1.0 limitation, that no if-statements are supported. That
limitation is no longer relevant for the module since the
corresponding code was changed with the last commit.

Additionally the type of two local variables were changed to
boolean (in the template that matches non-empty-text-elements) .