Minor changes module STLXML2EBU-TT
StartBox: The comment "but if two boxes are in the same line
(without style change), it puts a space between them" no longer
makes sense and was updated.

StartBox example modified by 072ad4b8: Was changed back to
contain three boxes (like example title/comment state).

EndBox: spanCreated is set to true() (if buffer not empty), as
buffer is written.