Minor change in calculation of line numbers
In STLXML2EBU-TT.xslt the number of subtitle lines is calculated
for the correct vertical positioning. When the line height is
DoubleHeight, the algorithm was limited to two subtitle lines
(further lines were ignored).

This commit makes a minor change in this calculation in such a
way that more than two subtitle lines will be detected correctly,
for single height and double height.

There are a few limitations in the algorithm:
* In a TextField a consistent line height is expected for all
  subtitle lines.
* Empty lines between two subtitle lines are not supported and
  may lead to an unexpected vertical positioning.
* A newline-Element at the end of the subtitle block (meaning no
  further subtitle line is following) may lead to an unexpected
  vertical positioning.