STLXML2EBU-TT: fix first word in TF ignored
Usually a subtitle is enclosed by StartBox/EndBox chars or at least
preceded by a DoubleHeight char i.e. the first child of STL's TF field
is an element. If this was not the case i.e. if the very first child of
the TF field was not an element but a text node, this text was ignored
by mistake.

Fixes IRT-Open-Source/scf#58.