STLXML-SplitBlocks: fix spread composite sequence
When in STLXML a character with diacritical chars is represented with
two Unicode codepoints instead of a single one (e.g. `J́`, which cannot
be represented with a single codepoint), it can occur that this
composite sequence is located at such a bad position that the present
module will move the two codepoints to different TTI blocks.

However this will contradict the correct later conversion from STLXML to
STL, as in Unicode the combining diacritical mark is a suffix. In STL in
contrast, the corresponding diacritical mark is a prefix. While
`STLXML2STL` swaps the order of such two codepoints, it cannot do in the
described case, as this only affects a single TTI block, but does not go
beyond TTI block borders.

So such composite sequence must not be splitted among different TTI
blocks.

Fixes IRT-Open-Source/scf#59.