STLXML2STL: fix composite sequence mapping in TF
Composite sequences with diacritical characters were not correctly
mapped from STLXML to STL in case of decomposed characters. Due to the
different order of graphic character and diacritical character in
STL and in Unicode, swapping the two (decomposed) characters is
required.

The necessary swap was already implemented for precomposed characters.
However so far for decomposed characters (e.g. `m̃`, which is not
available as precomposed character) it wasn't.

Fixes IRT-Open-Source/scf#55.