Service: disable EBU-TT-D(-Basic-DE) to STLXML/STL
If the source format EBU-TT-D or EBU-TT-D-Basic-DE was used with STLXML
or STL as target format, the conversion aborted with the error message
"map(*) expected, empty sequence found.", as specific source/target
format combinations were not considered in the code. As this conversion
is currently not supported, it is disabled now.

Also refactors the retrieval of the next format for easier usage.

Fixes IRT-Open-Source/scf#77.